requirements.txt
# Core RAG Frameworks
# NOTE: chromadb (Chroma) currently has compatibility issues with Python 3.14+.
#       Use Python 3.13 or earlier when installing these requirements.
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0

# LlamaIndex (for indexing strategies module)
llama-index>=0.12.0
llama-index-embeddings-openai
llama-index-llms-openai

# Vector Stores
faiss-cpu>=1.9.0
chromadb>=0.5.0

# Evaluation Metrics
rouge-score>=0.1.2
nltk>=3.8.0
scipy>=1.17.0

# LLM and Embeddings APIs
openai>=1.50.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
scikit-learn>=1.3.0
seaborn>=0.13.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# OpenTelemetry (for chromadb compatibility)
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0
